pylint corrections: Fix all the errors `make check-pylint` was showing
`make check-pylint` was launching quite a few errors. "pylint" is a tool
that should be used as a guide, not as a rule to follow, so some of those
errors can be ignored.

The ones that I'm going to ignore by default have been added in
`pylint.conf`, whereas some very special cases will need to be locally
disabled (with in-line comments).