tests: Add a small git repository and a first end-to-end test
The end-to-end tests will be stored in bash files following the convention:
1) Their file names will be in the form "tests-*.sh"
2) The tests will be declared as functions with names "test_*()"

The test added in this commit covers the test case where the branch is
specified, but no other "main branch reference" is. In that case, we
assume that the user wants to walk through all the entire branch, starting
from the first commit in the repository.

When adding this test, I've realised that the way I have implemented the
script is not the more intuitive one. What it currently does is assume that
the user wants to walk through the branch until the first commit that is
*not* contained in master. That's exactly the opposite of what it should
do. This needs to be fixed.