test/tests-branch-option.sh: New test for valid base branch
This is the ideal situation, with no error. The new option "--from-branch"
specifies what is the root branch which our branch of interest was forked
from.

In order to have this option well tested, we'll need error test cases like
"branch does not exist" or "--from-branch is not contained in --branch".